AssetSelect2 -> AssetSelect
